+++ common/src/main/java/com/illusivesoulworks/elytraslot/ElytraSlotConstants.java
@@ -22,7 +22,7 @@ import org.slf4j.LoggerFactory;
 
 public class ElytraSlotConstants {
 
-	public static final String MOD_ID = "elytraslot";
-	public static final String MOD_NAME = "Elytra Slot";
-	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+  public static final String MOD_ID = "elytraslot";
+  public static final String MOD_NAME = "Elytra Slot";
+  public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
 }