+++ common/src/main/java/com/illusivesoulworks/elytraslot/client/ElytraSlotLayer.java
@@ -18,6 +18,7 @@
 package com.illusivesoulworks.elytraslot.client;
 
 import com.illusivesoulworks.elytraslot.ElytraSlotCommonMod;
+import com.illusivesoulworks.elytraslot.platform.Services;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import java.awt.Color;
@@ -36,6 +37,7 @@ import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.PlayerModelPart;
+import net.minecraft.world.item.ArmorItem;
 
 public class ElytraSlotLayer<T extends LivingEntity, M extends EntityModel<T>>
     extends RenderLayer<T, M> {
@@ -53,15 +55,27 @@ public class ElytraSlotLayer<T extends LivingEntity, M extends EntityModel<T>>
     ElytraSlotCommonMod.getElytraRender(livingEntity).ifPresent(elytra -> {
       ResourceLocation resourcelocation;
 
+      if (elytra.stack().getItem() instanceof ArmorItem) {
+        return;
+      }
+
       if (livingEntity instanceof AbstractClientPlayer abstractclientplayer) {
 
         if (abstractclientplayer.isElytraLoaded() &&
             abstractclientplayer.getElytraTextureLocation() != null) {
           resourcelocation = abstractclientplayer.getElytraTextureLocation();
-        } else if (abstractclientplayer.isCapeLoaded() &&
-            abstractclientplayer.getCloakTextureLocation() != null &&
-            abstractclientplayer.isModelPartShown(PlayerModelPart.CAPE)) {
-          resourcelocation = abstractclientplayer.getCloakTextureLocation();
+        } else if (elytra.useCapeTexture()) {
+
+          if (Services.PLATFORM.isModLoaded("minecraftcapes") &&
+              Services.CLIENT.hasCustomCape(abstractclientplayer)) {
+            resourcelocation = Services.CLIENT.getCustomCape(abstractclientplayer);
+          } else if (abstractclientplayer.isCapeLoaded() &&
+              abstractclientplayer.getCloakTextureLocation() != null &&
+              abstractclientplayer.isModelPartShown(PlayerModelPart.CAPE)) {
+            resourcelocation = abstractclientplayer.getCloakTextureLocation();
+          } else {
+            resourcelocation = elytra.texture();
+          }
         } else {
           resourcelocation = elytra.texture();
         }