+++ common/src/main/java/com/illusivesoulworks/elytraslot/platform/services/IClientPlatform.java
@@ -15,12 +15,14 @@
  * License along with Elytra Slot. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.illusivesoulworks.elytraslot.client;
+package com.illusivesoulworks.elytraslot.platform.services;
 
-import java.awt.Color;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.entity.player.Player;
 
-public record ElytraRenderResult(Color color, ResourceLocation texture, boolean enchanted,
-                                 ItemStack stack, boolean useCapeTexture) {
+public interface IClientPlatform {
+
+  boolean hasCustomCape(Player player);
+
+  ResourceLocation getCustomCape(Player player);
 }