+++ fabric/src/main/java/com/illusivesoulworks/elytraslot/platform/FabricClientPlatform.java
@@ -15,12 +15,21 @@
  * License along with Elytra Slot. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.illusivesoulworks.elytraslot.client;
+package com.illusivesoulworks.elytraslot.platform;
 
-import java.awt.Color;
+import com.illusivesoulworks.elytraslot.platform.services.IClientPlatform;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.entity.player.Player;
 
-public record ElytraRenderResult(Color color, ResourceLocation texture, boolean enchanted,
-                                 ItemStack stack, boolean useCapeTexture) {
+public class FabricClientPlatform implements IClientPlatform {
+
+  @Override
+  public boolean hasCustomCape(Player player) {
+    return false;
+  }
+
+  @Override
+  public ResourceLocation getCustomCape(Player player) {
+    return null;
+  }
 }