+++ forge/src/main/java/com/illusivesoulworks/elytraslot/ElytraSlotForgeClientMod.java
@@ -47,9 +47,12 @@ public class ElytraSlotForgeClientMod {
   @SuppressWarnings({"rawtypes", "unchecked"})
   private static void addPlayerLayer(EntityRenderersEvent.AddLayers evt, String skin) {
     EntityRenderer<? extends Player> renderer = evt.getSkin(skin);
+    boolean slim = skin.equals("slim");
 
     if (renderer instanceof LivingEntityRenderer livingRenderer) {
       livingRenderer.addLayer(new ElytraSlotLayer(livingRenderer, evt.getEntityModels()));
+      livingRenderer.addLayer(
+          new ElytraSlotArmorLayer(livingRenderer, evt.getEntityModels(), slim));
     }
   }
 
@@ -59,6 +62,7 @@ public class ElytraSlotForgeClientMod {
 
     if (renderer != null) {
       renderer.addLayer(new ElytraSlotLayer<>(renderer, evt.getEntityModels()));
+      renderer.addLayer(new ElytraSlotArmorLayer<>(renderer, evt.getEntityModels(), false));
     }
   }