+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 Prior to version 6.0.0, this project used MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [6.1.0+1.19.2] - 2023.02.01
+### Added
+- Added Lil' Wings compatibility [#62](https://github.com/illusivesoulworks/elytraslot/issues/62)
+- Added Enderite Mod compatibility [#59](https://github.com/illusivesoulworks/elytraslot/issues/59)
+- Added Deeper and Darker compatibility [#57](https://github.com/illusivesoulworks/elytraslot/issues/57)
+- Added MinecraftCapes compatibility [#48](https://github.com/illusivesoulworks/elytraslot/issues/48)
+- Added Mekanism compatibility [#44](https://github.com/illusivesoulworks/elytraslot/issues/44)
+- Added Alex's Mobs compatibility [#28](https://github.com/illusivesoulworks/elytraslot/issues/28)
+
 ## [6.0.0+1.19.2] - 2022.08.11
 ### Changed
 - Updated to Minecraft 1.19.2